move static helper code to separate Helper class
