+++ src/main/java/com/cleanroommc/groovyscript/compat/mods/techreborn/DistillationTower.java
@@ -85,7 +85,7 @@ public class DistillationTower extends AbstractGenericTechRebornRegistry {
             ItemStack output2 = output.size() >= 2 ? output.get(1) : null;
             ItemStack output3 = output.size() >= 3 ? output.get(2) : null;
             ItemStack output4 = output.size() >= 4 ? output.get(3) : null;
-            DistillationTowerRecipe recipe = new DistillationTowerRecipe(TechReborn.getStackFromIIngredient(input.get(0)), input.size() >= 2 ? TechReborn.getStackFromIIngredient(input.get(1)) : null, output.get(0), output2, output3, output4, time, perTick, oreDict);
+            DistillationTowerRecipe recipe = new DistillationTowerRecipe(Helper.getStackFromIIngredient(input.get(0)), input.size() >= 2 ? Helper.getStackFromIIngredient(input.get(1)) : null, output.get(0), output2, output3, output4, time, perTick, oreDict);
             ModSupport.TECH_REBORN.get().distillationTower.add(recipe);
             return recipe;
         }
