+++ src/main/java/com/cleanroommc/groovyscript/compat/mods/techreborn/IndustrialGrinder.java
@@ -99,7 +99,7 @@ public class IndustrialGrinder extends AbstractGenericTechRebornRegistry {
             ItemStack output2 = output.size() >= 2 ? output.get(1) : null;
             ItemStack output3 = output.size() >= 3 ? output.get(2) : null;
             ItemStack output4 = output.size() >= 4 ? output.get(3) : null;
-            IndustrialGrinderRecipe recipe = new IndustrialGrinderRecipe(TechReborn.getStackFromIIngredient(input.get(0)), fluidInput.get(0), output.get(0), output2, output3, output4, time, perTick, oreDict);
+            IndustrialGrinderRecipe recipe = new IndustrialGrinderRecipe(Helper.getStackFromIIngredient(input.get(0)), fluidInput.get(0), output.get(0), output2, output3, output4, time, perTick, oreDict);
             ModSupport.TECH_REBORN.get().industrialGrinder.add(recipe);
             return recipe;
         }
