+++ src/main/java/com/cleanroommc/groovyscript/compat/mods/techreborn/PlateBendingMachine.java
@@ -80,7 +80,7 @@ public class PlateBendingMachine extends AbstractPraescriptumRegistry {
             if (!validate()) return null;
 
             Recipe recipe = Recipes.plateBendingMachine.createRecipe();
-            recipe.withInput(input.stream().map(TechReborn::toInputIngredient).collect(Collectors.toList()));
+            recipe.withInput(input.stream().map(Helper::toInputIngredient).collect(Collectors.toList()));
             output.forEach(recipe::withOutput);
             recipe.withEnergyCostPerTick(perTick);
             recipe.withOperationDuration(time);
